feat: add the base template for the entire website
Amusingly, the content inside the <head> tag was complex and large
enough to deserve its own macro, which is basically a better POSIX sh
function, or a procedure. The macros inside a macro file can be called
upon by importing the macro file at the top of the tera template.

The header, nav, and the footer child elements of the body are defined
in the base template because they are replicated on all possible pages
on the theme. The main tag, however, isn't and its content varies across
pages, which is why it hasn't been defined in the base template.

Nevertheless, most of the content inside header, nav, and footer can be
overridden because most of the content is wrapped inside blocks. A
custom template which extends base can be override specific blocks to
embed custom content. In addition, much of the content has been made
optional using if-else expressions which depend on the values defined in
config.toml.