feat: add the macro for populating head tag
As mentioned in the commit that adds base.html, this macro substitutes
content that should be in the head tag.

I'm not confident if I've mentioned the relative paths correctly. Guess
we'll find out? I was also not able to figure out how to determine the
feed type of a custom feed name, whether its atom or rss.

The self keyword addresses macro functions defined in the same file. The
`self::keywords()` snipppet means that the `keywords()` macro defined
in the same file is being referenced.