feat: add index.html, the home page of the theme
This page uses base.html as its foundation and overwrites the block for
the main tag. Just how we used macros in base.html, we've used a macro
for embedding the contents of an article because this will be used in
several places. Yes, deciding the layout and structure of a website
needs to be done before you start writing templates for it.

The macros used in Tera support passing arguments, unlike bash
functions (which are procedures rather than functions). Since the home
page will have all the articles/posts we've written, we loop through the
`section.pages` variable to get all the articles inside the root
section. Each article itself is denoted by the `post` variable and this
variable is passed as an argument to the article macro functions which
use the values in post to create the content that we want.