feat: add template for a 404 page
The 404 page doesn't need to have meta information such as the opengraph
tags. In addition, it won't have any canonical URL, which is the
`base_canonical` block has been overwritten to include only the `base`
HTML tag.

The titles have been overwritten with `404`. I tried to be a bit
creative with the 404 error message. :sweat_smile: