feat: add the CSS for the kangae theme
Phew, creating this CSS file took a significant amount of time,
primarily because I'm not a web developer and I have never used CSS
before for any project.

Yes, the size of this CSS file is larger than it should be because of
comments and line breaks intended for readability. I firmly believe that
trading off a few extra kilobytes is worth it for people who want to
understand how the CSS of this website is designed by clicking
view-source in his web browser. Understanding the HTML or CSS of a
website shouldn't require someone to use prettier or other formatters to
undo the mess the minification makes to save a few kilobytes.

The step and space variables are taken from the https://utopia.fyi
website. These clamp() values enable fluid typography and space
distribution on the kangae theme.