style: improve APCA contrast value in light mode
The APCA contrast value was improved from 92.68 to 97.41 by using
`f8f8f8` as the background color instead of `f0f0f0` against text color
of `282828`. This helps us comply with APCA on all commonly used font
size and weigh combinations.

A note was added to the border color which highlights that the border
color is chosen deliberately to have a lower contrast than the fg-bg
combination.