feat: open external links in navbar in a newtab
If an external link is added in a navbar, it should have
`target="_blank" rel="noopener noreferrer"`.

Yeah, looks a bit messy. This was needed because if we specify `rel`
multiple times, only one is considered and the others are ignored. We
have to specify the value of `rel` in a single value assignment
operation and the values should be separated by spaces.